Display data cache info Shows filenames and cache file sizes

Display data cache info Shows filenames and cache file sizes

Usage

# S3 method for class 'hockeystick_cache_info'
print(x, ...)

Arguments

x

filenames

...

Additional parameters

Value

Display data cache info. No return value, called for side effect.